associate topics with quiz
