+++ app/models/quiz.rb
@@ -1,4 +1,6 @@
 class Quiz < ApplicationRecord
+  include TopicMatchable
+
   STATUS_ENUM = {
     pending: 0,
     active: 1,
@@ -29,6 +31,8 @@ class Quiz < ApplicationRecord
   has_many :quiz_questions, dependent: :destroy
   has_many :quiz_question_answers, through: :quiz_questions
   has_many :quiz_responses
+  has_many :citation_topics, foreign_key: :citation_id, primary_key: :citation_id
+  has_many :topics, through: :citation_topics
 
   validates_presence_of :citation_id
   validate :prompt_params_text_valid_json
